Moved exceptions from constructors.py to exceptions.py
