+++ impf/constructors.py
@@ -21,31 +21,6 @@ except ModuleNotFoundError:
     if settings.ZULIP_ENABLED: logger.warning('Zulip package not found, but ZULIP_ENABLED configured '
                                               '- cannot send alerts via zulip until package is installed')
 
-class AdvancedSessionError(Exception):
-    def __init__(self, code, message):
-        super().__init__(message)
-        self.code = code
-        self.message = message
-
-    def __repr__(self):
-        return f'AdvancedSessionError({self.code!r}, {self.message!r})'
-
-    def __str__(self):
-        return f'[{self.code}] {self.message}'
-
-class AdvancedSessionCache(Exception):
-    def __init__(self, code, message):
-        super().__init__(message)
-        self.code = code
-        self.message = message
-
-    def __repr__(self):
-        return f'AdvancedSessionCache({self.code!r}, {self.message!r})'
-
-    def __str__(self):
-        return f'Cookies likely expired – [{self.code}] {self.message}'
-
-
 def browser_options():
     """ Helper function to build Selenium Browser options """
     opts = Options()
